fix: corrige a penultima operacao ao alterar o tipo da variavel de int para decimal

+++ Calculadora/Calculadora.cs
@@ -3,28 +3,28 @@ namespace Calculadora
 {
     public class Calculadora
     {
-        
+
         public Operacoes calcular(Operacoes operacao)
         {
-            switch(operacao.operador)
+            switch (operacao.operador)
             {
-                case '+': operacao.resultado= soma(operacao);break;
-                case '-': operacao.resultado = subtracao(operacao);break;
-                case '*': operacao.resultado = multiplicacao(operacao);break;
+                case '+': operacao.resultado = soma(operacao); break;
+                case '-': operacao.resultado = subtracao(operacao); break;
+                case '*': operacao.resultado = multiplicacao(operacao); break;
                 case '/': operacao.resultado = divisao(operacao); break;
                 default: operacao.resultado = 0; break;
             }
             return operacao;
         }
-        public int soma(Operacoes operacao)
+        public decimal soma(Operacoes operacao)
         {
             return operacao.valorA + operacao.valorB;
         }
-        public int subtracao(Operacoes operacao)
+        public decimal subtracao(Operacoes operacao)
         {
             return operacao.valorA - operacao.valorB;
         }
-        public int multiplicacao(Operacoes operacao)
+        public decimal multiplicacao(Operacoes operacao)
         {
             return operacao.valorA * operacao.valorB;
         }
